NAB statement file (NAI/BAI2), parsing logic
classes for parsing a NAB "Account Information File (NAI/BAI2)". the
format here is a rather baroque mashup of a CSV restricted to 80 chars
(the lineage comes from punchcards) with continuation records, so the
parser makes available methods to simplify this

there's not too much other complexity here, although we shall see once
we start using this in anger. add test coverage around all of the above